requirements.txt
git-python
async-generator
attrs
bcrypt
certifi
cffi
coloredlogs~=15.0.1
cryptography
cycler
fonttools
h11
humanfriendly
idna
interchange
jarowinkler
kiwisolver
Levenshtein~=0.18.1
lxml~=4.8.0
matplotlib~=3.5.1
monotonic
networkx~=2.6.3
numpy~=1.21.6
outcome
packaging
pandas~=1.3.5
pansi
paramiko~=2.10.3
Pillow
py2neo~=2021.2.3
pycparser
Pygments
PyNaCl
pyOpenSSL
pyparsing
pyreadline
pyreadline3
PySocks
python-dateutil
pytz
rapidfuzz
ruamel.yaml
ruamel.yaml.clib
selenium~=3.141.0
six
sniffio
sortedcontainers
trio
trio-websocket
typing_extensions
urllib3
wsproto
tqdm~=4.64.0
bitarray~=2.4.1
requests